Delete certificate model and reorganize components
Removed the certificate model since it's no longer in use. Renamed and moved several modules into a new `components` directory for better organization. Updated the project name in `pyproject.toml` and CLI commands accordingly.
