+++ device_certificate_report/components/report_generation.py
@@ -9,7 +9,6 @@ from reportlab.platypus import (
     Table,
     TableStyle,
     Image,
-    PageBreak,
 )
 from reportlab.lib.styles import getSampleStyleSheet
 from reportlab.lib.pagesizes import letter
