+++ pyproject.toml
@@ -28,4 +28,4 @@ requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
-device-certificate-report = "device_certificate_report.main:app"
+panos-device-certificate-report = "device_certificate_report.main:app"
